Add persistent config using configstore
On Linux, config goes in ~/.config/tagtime-desktop/config.json.
On Windows I think it will go in $USER/.config/...

Also initialise user's seed to a random value on first run.

Tests modify the user config and then restore it. There's probably
a better way to do this (mock the config?), but it works.