Add coverage of (main process) end to end tests
gulp is used to build an istanbul-instrumented version of the app, which
is run by the end to end tests instead of the usual version.

The main process saves coverage information to a path specified via an
environment variable. The gulp report:e2e task then merges the coverage
files into a coverage-final.json which forms part of the normal input to
the existing report generation.